[metasrv] refactor: move raft-store impl into dir store/
