+++ metasrv/src/lib.rs
@@ -27,3 +27,4 @@ pub mod errors;
 pub mod executor;
 pub mod meta_service;
 pub mod metrics;
+mod store;
