+++ metasrv/src/meta_service/meta_leader.rs
@@ -22,15 +22,15 @@ use common_meta_types::Node;
 use common_meta_types::NodeId;
 use common_tracing::tracing;
 
+use crate::errors::ForwardToLeader;
+use crate::errors::InvalidMembership;
+use crate::errors::MetaError;
+use crate::errors::RetryableError;
 use crate::meta_service::message::AdminRequest;
 use crate::meta_service::message::AdminResponse;
 use crate::meta_service::AdminRequestInner;
-use crate::meta_service::ForwardToLeader;
-use crate::meta_service::InvalidMembership;
 use crate::meta_service::JoinRequest;
-use crate::meta_service::MetaError;
 use crate::meta_service::MetaNode;
-use crate::meta_service::RetryableError;
 
 /// The container of APIs of a metasrv leader in a metasrv cluster.
 ///
