+++ metasrv/src/meta_service/raftmeta_test.rs
@@ -31,11 +31,11 @@ use maplit::btreeset;
 use pretty_assertions::assert_eq;
 
 use crate::configs;
+use crate::errors::RetryableError;
 use crate::meta_service::message::AdminRequest;
 use crate::meta_service::AdminRequestInner;
 use crate::meta_service::JoinRequest;
 use crate::meta_service::MetaNode;
-use crate::meta_service::RetryableError;
 use crate::proto::meta_service_client::MetaServiceClient;
 use crate::tests::assert_meta_connection;
 use crate::tests::service::new_test_context;
